Conflicts:
	AI/Interfaces/Java/CMakeLists.txt
	AI/Wrappers/LegacyCpp/AIAICallback.cpp
	installer/spring.nsi
	rts/ExternalAI/Interface/AISCommands.h
	rts/ExternalAI/Interface/SSkirmishAICallback.h
	rts/ExternalAI/SkirmishAILibrary.h

+++ installer/make_source_package.sh
@@ -78,7 +78,6 @@ include=" \
  $dir/tools/DedicatedServer/"
 
 # On linux, win32 executables are useless.
-# TASClient is windows only.
 exclude_from_all=""
 linux_exclude="${exclude_from_all}
 	${dir}/installer/include/
@@ -91,7 +90,6 @@ linux_exclude="${exclude_from_all}
 	${dir}/installer/*.nsh
 	${dir}/installer/make_installer.pl
 	${dir}/installer/make_luaui_nsh.py
-	${dir}/installer/tasclient_download.sh
 	${dir}/installer/springlobby_download.sh"
 linux_include=""
 windows_exclude="${exclude_from_all}"
@@ -106,31 +104,34 @@ git submodule update --init
 cd ..
 [ -n "$linux_exclude" ] && rm -rf $linux_exclude
 [ -n "$lzma" ] && echo "Creating .tar.lzma archive ($lzma)" && \
-	tar --lzma -c -f "../$lzma" $include $linux_include
+	tar --lzma -c -f "../$lzma" $include $linux_include --exclude=.git
 [ -n "$tbz" ] && echo "Creating .tar.bz2 archive ($tbz)" && \
-	tar -c -j -f "../$tbz" $include $linux_include
+	tar -c -j -f "../$tbz" $include $linux_include --exclude=.git
 [ -n "$tgz" ] && echo "Creating .tar.gz archive ($tgz)" && \
-	tar -c -z -f "../$tgz" $include $linux_include
+	tar -c -z -f "../$tgz" $include $linux_include --exclude=.git
 cd ..
 echo 'Cleaning'
 rm -rf lf
 
-### TODO: needs fixing (not really using CRLF)
-# Windows line endings, .zip/.7z package
-#echo 'Exporting checkout dir with CRLF line endings'
-git clone -n . crlf/$dir
-cd crlf/$dir
-git config core.autocrlf true
-git checkout $branch
-git submodule update --init
-cd ..
+if [ -n "$zip" ] || [ -n "$seven_zip" ]; then
+	### TODO: needs fixing (not really using CRLF)
+	# Windows line endings, .zip/.7z package
+	#echo 'Exporting checkout dir with CRLF line endings'
+	git clone -n . crlf/$dir
+	cd crlf/$dir
+	git config core.autocrlf true
+	git checkout $branch
+	git submodule update --init
+	cd ..
 
-[ -n "$windows_exclude" ] && rm -rf $windows_exclude
-[ -n "$zip" ] && [ -x /usr/bin/zip ] && echo "Creating .zip archive ($zip)" && \
-	/usr/bin/zip -q -r -u -9 "../$zip" $include $windows_include
-[ -n "$seven_zip" ] && [ -x /usr/bin/7z ] && echo "Creating .7z archive ($seven_zip)" && \
-	/usr/bin/7z a -t7z -m0=lzma -mx=9 -mfb=64 -md=32m -ms=on "../$seven_zip" $include >/dev/null
-cd ..
-echo 'Cleaning'
-rm -rf crlf
+	[ -n "$windows_exclude" ] && rm -rf $windows_exclude
+	[ -n "$zip" ] && [ -x /usr/bin/zip ] && echo "Creating .zip archive ($zip)" && \
+		/usr/bin/zip -q -r -u -9 "../$zip" $include $windows_include
+	[ -n "$seven_zip" ] && [ -x /usr/bin/7z ] && echo "Creating .7z archive ($seven_zip)" && \
+		/usr/bin/7z a -t7z -m0=lzma -mx=9 -mfb=64 -md=32m -ms=on "../$seven_zip" $include >/dev/null
+	cd ..
+	echo 'Cleaning'
+	rm -rf crlf
+fi
+
 cd ..
